Grassroots Economics - Book
This book is not an answer, but an invitation—an invitation to explore new (and old) ways of coordinating resources, restoring trust, and creating economies that serve the well-being of all. I hope you enjoy it, share it, and most of all, find ways to bring its lessons into your own life.